Some improvements and fixes (#29)
* moved configuration to separate central file, added pushover notification service, bumped version

* fix missing timer restart

* updates regarding new esp8266 sdk (2.5.2)

* added auto download for vs1053 patches

* lot of changes

* move ifdef to correct position

+++ RfidShelf/ShelfWeb.h
@@ -12,8 +12,6 @@
 #include "ShelfRfid.h"
 #include "ShelfVersion.h"
 
-//extern SdFat SD;
-
 class ShelfWeb {
   public:
     ShelfWeb(ShelfPlayback &playback, ShelfRfid &rfid, SdFat &sd);
@@ -31,7 +29,7 @@ class ShelfWeb {
     SdFile _uploadFile;
     uint32_t _uploadStart;
     void returnOK();
-    void returnHttpStatus(uint8_t statusCode, String msg);
+    void returnHttpStatus(uint16_t statusCode, String msg);
     void renderDirectory(String &path);
     bool loadFromSdCard(String &path);
     void handleWriteRfid(String &folder);
